Rework ServerPollingJob to be TaskSpec
